Fixed #363.

+++ permission/src/main/java/com/yanzhenjie/permission/checker/StorageWriteTest.java
@@ -16,6 +16,7 @@
 package com.yanzhenjie.permission.checker;
 
 import android.os.Environment;
+import android.text.TextUtils;
 
 import java.io.File;
 
@@ -29,13 +30,18 @@ class StorageWriteTest implements PermissionTest {
 
     @Override
     public boolean test() throws Throwable {
+        if (!TextUtils.equals(Environment.MEDIA_MOUNTED, Environment.getExternalStorageState())) return true;
+
         File directory = Environment.getExternalStorageDirectory();
         if (!directory.exists() || !directory.canWrite()) return false;
         File parent = new File(directory, "Android");
         if (parent.exists() && parent.isFile()) if (!parent.delete()) return false;
         if (!parent.exists()) if (!parent.mkdirs()) return false;
         File file = new File(parent, "ANDROID.PERMISSION.TEST");
-        if (file.exists()) return file.delete();
-        else return file.createNewFile();
+        if (file.exists()) {
+            return file.delete();
+        } else {
+            return file.createNewFile();
+        }
     }
 }
